Compile for MacOS
With Linux, linking to the libm library is easy, but on MacOS, there
exists no keg in homebrew. Fix that by making CMake platorm specific for
loading the library for argp.

On Mac use, `brew install argp-standalone` to instal libargp, and use
`brew install readline` to install libreadline.

Lastly, update the RADME.md to show steps needed for compile.