WASM compilation
A minimal setup to compile for WASM. Some improvements that could be
made are that we create a separate library, and different targets for
normal compilation & separate target for WASM. This way, 1 program, many
targets. Ideally, we get a kind of argp-standalone version into the
repository to compile as a separate lib to include. This way, we don't
need special install requirements for Mac as well as for WASM. We'd just
need to get gnulib into this project, and pull out argp specifically for
separate compilation.