separate lexer & parser
Before we're able to make a LSP, we first need to separate lexing from
the parsing. This needs to be done because the LSP needs to have tokens
made available to it to know how to make a decision for any event.

This should've been done a long time ago, but now is better
late than never.