Attach offset and size info to tokens
LSP could send events with cursor positions in the middle of tokens, but
the server for the LSP would need to somehow backtrack to figure out what
token is at that position.

With meta added to the tokens, the server could just enumerate all the
tokens for a given line (which would be done anyway), and then loop
through the list of tokens to find which token the cursor is ontop of.